refactor: 패키지명에 오타가 난 것을 response로 변경
