+++ src/main/java/com/todo/post/dto/response/PostResponse.java
@@ -1,0 +1,4 @@
+package com.todo.post.dto.response;
+
+public class PostResponse {
+}
