+++ src/main/java/com/todo/post/service/PostService.java
@@ -1,7 +1,7 @@
 package com.todo.post.service;
 
 import com.todo.post.domain.Post;
-import com.todo.post.dto.reponse.UserWithPostsResponse;
+import com.todo.post.dto.response.UserWithPostsResponse;
 import com.todo.post.dto.request.PostCreateRequest;
 import com.todo.post.dto.request.PostUpdateRequest;
 import com.todo.post.repository.PostRepository;
